+++ server/src/api/controllers/ShortURL/RedirectShortUrl.controller.ts
@@ -1,12 +1,12 @@
-import IShortUrl from "../models/interfaces/IShortUrl";
-import { ShortUrls } from "../models/shortURL.model";
+import IShortUrl from "../../models/interfaces/IShortUrl";
+import { ShortUrls } from "../../models/shortURL.model";
 import { Response } from "express";
-import IRequest from "./interfaces/IRequest";
-import IReqParams from "./interfaces/IReqParams";
-import IMessage from "./interfaces/IMessage";
-import IResponse from "./interfaces/IResponse";
-import ErrorHandler from "../middleware/Handler/ErrorHandler";
-import HttpException from "../middleware/Exceptions/HttpException";
+import IRequest from "../interfaces/IRequest";
+import IReqParams from "../interfaces/IReqParams";
+import IMessage from "../interfaces/IMessage";
+import IResponse from "../interfaces/IResponse";
+import ErrorHandler from "../../middleware/Handler/ErrorHandler";
+import HttpException from "../../middleware/Exceptions/HttpException";
 
 const RedirectShortUrl = async (req: IRequest, res: Response) => {
   const msg: IMessage = { isSuccessful: false, message: "", info: null };