+++ server/src/api/routes/ShortURL.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import TestApi from "../controllers/Test.controller";
-import CreateShortUrl from "../controllers/CreateShortUrl.controller";
-import RedirectShortUrl from "../controllers/RedirectShortUrl.controller";
+import CreateShortUrl from "../controllers/ShortURL/CreateShortUrl.controller";
+import RedirectShortUrl from "../controllers/ShortURL/RedirectShortUrl.controller";
 
 const ShortUrlRoutes: Router = Router();